feat(web): Refactored various planner.vue only functions into their usages, partial #322
